feat: Implement human support API endpoints for detailed monitoring
- Added v1 and v2 URL configurations for human support detailed monitoring.
- Created views for detailed monitoring, including ongoing, awaiting, agents, and status for v1.
- Introduced a new view for agents in v2.
- Established necessary permissions and query parameter handling for project-specific data retrieval.
